backend/requirements.txt
fastapi
uvicorn
google-generativeai
gTTS
requests
python-dotenv
python-multipart
